+++ lib/sensitive_data_filter/middleware/occurrence.rb
@@ -18,12 +18,20 @@ module SensitiveDataFilter
         @original_env_parser.ip
       end
 
-      def original_params
-        @original_env_parser.params
+      def original_query_params
+        @original_env_parser.query_params
       end
 
-      def filtered_params
-        @filtered_env_parser.params
+      def original_body_params
+        @original_env_parser.body_params
+      end
+
+      def filtered_query_params
+        @filtered_env_parser.query_params
+      end
+
+      def filtered_body_params
+        @filtered_env_parser.body_params
       end
 
       def_delegators :@original_env_parser, :request_method, :url, :session
@@ -34,12 +42,13 @@ module SensitiveDataFilter
 
       def to_h
         {
-          origin_ip:       origin_ip,
-          request_method:  request_method,
-          url:             url,
-          filtered_params: filtered_params,
-          session:         session,
-          matches_count:   matches_count
+          origin_ip:             origin_ip,
+          request_method:        request_method,
+          url:                   url,
+          filtered_query_params: filtered_query_params,
+          filtered_body_params:  filtered_body_params,
+          session:               session,
+          matches_count:         matches_count
         }
       end